Removed tests from main directories.
