+++ EMRio.py
@@ -36,30 +36,13 @@
 #		instancegroup objects. More info can be found here:
 # 		boto.cloudhackers.com/en/latest/ref/emr.html#boto.emr.emrobject.JobFlow
 
-import datetime
-import json
 import sys
-from datetime import timedelta
 from optparse import OptionParser
-
-from iso8601 import parse_date
-
-# The user doesn't necessarily need mrjob, so we check to see if
-# they have mrjob, if not, then we use the backup BOTO
-import boto.exception
-try:
-	mrjob = True
-	from mrjob.emr import EMRJobRunner
-except ImportError:
-	mrjob = None
-
-	from boto.emr.connection import EmrConnection
-
 
 from ec2_cost import EC2
 from simulate_jobs import Simulator
 from graph_jobs import cost_graph, total_hours_graph
-from job_filter import JobFilter
+from job_filter import get_job_flows
 from optimizer import convert_to_yearly_hours, Optimizer
 
 
@@ -130,30 +113,6 @@ def make_option_parser():
 		'-g', '--graph', dest='graph', type='string', default='None',
 		help='Load a graph of the job flows. Current graphs are: cost, total_usage')
 	return option_parser
-
-
-def get_job_flows(options):
-	"""This will check the options and use a file if provided or
-	will get the job_flow data from amazon's cluster if no file
-	is provided
-
-	Args:
-		options: An OptionParser object that has args stored in it.
-
-	Returns:
-		job_flows: A list of dicts of jobs that have run over a period of time.
-	"""
-	job_filter = JobFilter(options)
-	job_flows = []
-	if(options.file_inputs):
-		job_flows = handle_job_flows_file(options.file_inputs)
-	else:
-		job_flows = get_job_flow_objects(options)
-	job_flows = job_filter.filter_jobs(job_flows)
-
-	# sort job flows before running simulations.
-	job_flows = sorted(job_flows, cmp=sort_by_job_times)
-	return job_flows
 
 
 def get_best_instance_pool(options, job_flows):
@@ -238,50 +197,6 @@ def simulate_job_flows(job_flows, pool):
 	return optimal_logged_hours, demand_logged_hours
 
 
-def handle_job_flows_file(filename):
-	"""If you specify a file of job_flow objects, this function loads them. Will
-	try comma-separated JSON objects and per-line objects before failing.
-	"""
-	try:
-		current_file = open(filename, 'r')
-		contents = current_file.read().rstrip('\n')[:-1]
-		job_flows = json.loads(contents)
-		return job_flows
-	except ValueError:
-		pass  # Might be a different format..
-	job_flows = []
-	current_file = open(filename, 'r')
-	for line in current_file.readlines():
-		job_flows.append(json.loads(line))
-	return job_flows
-
-
-def get_job_flows_from_amazon(options):
-	"""gets all the job flows from amazon and converts them into
-	a dict for compatability with loading from a file
-	"""
-	now = datetime.datetime.utcnow()
-	job_flows = get_job_flows(options.conf_path, options.max_days_ago, now=now)
-	dict_job_flows = []
-	for job in job_flows:
-		job_dict = job.__dict__
-		new_list = []
-		for instance in job.instancegroups:
-			new_list.append(instance.__dict__)
-		job_dict['instancegroups'] = new_list
-		dict_job_flows.append(job_dict)
-	job_flows = dict_job_flows
-	return job_flows
-
-
-def sort_by_job_times(job1, job2):
-	"""Sorting comparator for job_flow objects"""
-	date1 = job1.get('startdatetime')
-	date2 = job2.get('startdatetime')
-	return_time = -1 if date1 < date2 else 1
-	return return_time
-
-
 def output_statistics(log, pool, demand_log,):
 	"""Once everything is calculated, output here"""
 
@@ -310,94 +225,6 @@ def output_statistics(log, pool, demand_log,):
 	print "Cost for Reserved Instance: ", optimized_cost
 	print "Cost for all On-Demand: ", demand_cost
 	print "Money Saved: ", (demand_cost - optimized_cost)
-
-
-#### Stuff taken from mrjob.tools.emr.audit_usage #####
-def get_job_flow_objects(conf_path, max_days_ago=None, now=None):
-	"""Get relevant job flow information from EMR.
-
-	:param str conf_path: Alternate path to read :py:mod:`mrjob.conf` from, or
-						``False`` to ignore all config files.
-	:param float max_days_ago: If set, don't fetch job flows created longer
-								than this many days ago.
-	:param now: the current UTC time, as a :py:class:`datetime.datetime`.
-				Defaults to the current time.
-	"""
-	if now is None:
-		now = datetime.utcnow()
-	emr_conn = None
-	if mrjob:
-		emr_conn = EMRJobRunner(conf_path=conf_path).make_emr_conn()
-	else:
-		emr_conn = EmrConnection()
-	# if --max-days-ago is set, only look at recent jobs
-	created_after = None
-	if max_days_ago is not None:
-		created_after = now - timedelta(days=max_days_ago)
-
-	return describe_all_job_flows(emr_conn, created_after=created_after)
-
-
-def describe_all_job_flows(emr_conn, states=None, jobflow_ids=None,
-							created_after=None, created_before=None):
-	"""Iteratively call ``EmrConnection.describe_job_flows()`` until we really
-	get all the available job flow information. Currently, 2 months of data
-	is available through the EMR API.
-
-	This is a way of getting around the limits of the API, both on number
-	of job flows returned, and how far back in time we can go.
-
-	:type states: list
-	:param states: A list of strings with job flow states wanted
-	:type jobflow_ids: list
-	:param jobflow_ids: A list of job flow IDs
-	:type created_after: datetime
-	:param created_after: Bound on job flow creation time
-	:type created_before: datetime
-	:param created_before: Bound on job flow creation time
-	"""
-	all_job_flows = []
-	ids_seen = set()
-
-	if not (states or jobflow_ids or created_after or created_before):
-		created_before = datetime.datetime.utcnow() + timedelta(days=1)
-
-	while True:
-		if created_before and created_after and created_before < created_after:
-			break
-
-		try:
-			results = emr_conn.describe_jobflows(
-				states=states, jobflow_ids=jobflow_ids,
-				created_after=created_after, created_before=created_before)
-		except boto.exception.BotoServerError, ex:
-			if 'ValidationError' in ex.body:
-
-				break
-			else:
-				raise
-
-		# don't count the same job flow twice
-		job_flows = [jf for jf in results if jf.jobflowid not in ids_seen]
-
-		all_job_flows.extend(job_flows)
-		ids_seen.update(jf.jobflowid for jf in job_flows)
-
-		if job_flows:
-			# set created_before to be just after the start time of
-			# the first job returned, to deal with job flows started
-			# in the same second
-			min_create_time = min(parse_date(jf.creationdatetime)
-									for jf in job_flows)
-			created_before = min_create_time + timedelta(seconds=1)
-			# if someone managed to start 501 job flows in the same second,
-			# they are still screwed (the EMR API only returns up to 500),
-			# but this seems unlikely. :)
-		else:
-			if not created_before:
-				created_before = datetime.utcnow()
-			created_before -= timedelta(weeks=2)
-	return all_job_flows
 
 
 if __name__ == '__main__':
